// src/app/languages/interface/more/track.interface.ts
export interface ITrack {  
    trackOrder: string; 
    orderId: string;
    items: string;
    total: string;
    estimatedDeliveryon: string;
    orderPlaced: string;
    orderPacked: string;
    orderShipped: string;
    outForDelivery: string;
    delivered: string;
    deliveryAddress: string;
    orderprocessing: string;
    sellerpacked: string;
    sellerhanded: string;  
    courierisout: string;
    itemdelivered: string;
    rateproduct: string;
    
 }